requirements.txt
pydantic==2.5.2
pyyaml